folium TimeSliderChoropleth with nasa srdb - sort of working for one site in North Wales
Files:
  - time_series_choro.py :: standalone python or run as vscode notebook
  - time_series_choro.html :: output by `python3  time_series_choro.py`

For debug purposes, just displays one site near North Wales coast. This is
the site that has the most number of timestamps (13 in total)
You need to zoom in alot to see the coloured square on the folium map.

Note - the timeslider is slighlty out of order. It starts  2002 to 2016
and then 1999 to 2001